pyproject.toml
[project]
name = "causal-armor-test-harness"
version = "0.1.0"
description = "AgentDojo benchmark harness for CausalArmor prompt-injection defense"
requires-python = ">=3.11"
dependencies = [
    "agentdojo>=0.1.30",
    "langgraph>=0.2",
    "langchain-google-genai>=2.0",
    "langchain-core>=0.3",
    "causal-armor[gemini] @ git+https://github.com/prashantkul/causal-armor.git",
    "causal-armor-langgraph-demo @ git+https://github.com/prashantkul/causal-armor-langgraph-demo.git",
    "typer>=0.12",
    "rich>=13.0",
    "python-dotenv>=1.0",
]

[project.scripts]
ca-bench = "harness.cli:app"

[project.optional-dependencies]
dev = ["pytest>=8.0", "pytest-asyncio>=0.23"]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/harness"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
